+++ generate.js
@@ -18,7 +18,10 @@ var options = require('optimist')
 			  .alias('t', 'standuptime')
 			  .describe('t', 'Standup meeting time; if not defined, split time is midnight')
 			  .alias('n', 'name')
-			  .describe('n', 'Name of the sprint');
+			  .describe('n', 'Name of the sprint')
+			  .alias('s', 'save')
+			  .describe('s', 'Save configuration')
+			  .boolean('s');
 
 var optionArgs = options.argv;
 var path = require('path');
@@ -26,16 +29,17 @@ var path = require('path');
 global.settings = require('./settings');
 settings.root   = __dirname.replace(/\/+$/, "");
 settings.exportPath = path.join(settings.root, 'export');
+settings.configPath = path.join(settings.root, 'config');
 
-var CardReceiver = require('./lib/cardreceiver');
-var CardStatistics = require('./lib/cardstatistics');
+var Helper = require('./lib/helper');
 
 var lists = optionArgs.l;
 var days = optionArgs.d;
 var resources = optionArgs.r;
 var finishedList = optionArgs.f;
-var standuptime = optionArgs.t;
+var standupTime = optionArgs.t;
 var name = optionArgs.n;
+var save = optionArgs.s;
 
 if (!lists || !lists.length) 
 {
@@ -69,58 +73,11 @@ if (!finishedList) {
 days = days.split(',');
 resources = resources.split(',');
 
-start();
-
-function start() {
-	var receiver = new CardReceiver(settings.applicationKey, settings.userToken, settings.boardId);
-	receiver.receive(lists.split(','), function(err, cards) {
-		if (err) {
-			console.log(err instanceof Error ? err.message : err);
-			return;
-		}
-
-		if (cards.length > 0) {
-			var statistics = new CardStatistics();
-			statistics.generate(cards, finishedList, standuptime, function(err, data) {
-
-				if (err) {
-					console.log(err);
-					return;
-				}
-
-				printStatistics(data);
-
-				statistics.export(data, resources, days, name, function(err) {
-					if (err) {
-						console.log(err);
-						return;
-					}
-
-					console.log("Exported successfully");
-				})				
-
-			});
-		} else {
-			console.log("No cards found.");
-		}
-	});
+var helper = new Helper();
+if (save) {
+	console.log("Generating configuration ...");
+	helper.saveConfiguration(lists, days, resources, finishedList, standupTime, name);
+} else {
+	console.log("Starting export ...");
+	helper.generateAndExport(lists, days, resources, finishedList, standupTime, name);
 }
-
-function printStatistics(data) {
-	console.log("");
-	console.log("Statistics");
-	console.log("----------");
-	console.log("Cards (total):    " + (data.cardsopen + data.cardsfinished));
-	console.log("Cards (open):     " + data.cardsopen);
-	console.log("Cards (finished): " + data.cardsfinished);
-	console.log("");
-	console.log("Estimate (total): " + data.estimate);
-	console.log("Estimate (open):  " + (data.estimate - data.estimatedone));
-	console.log("Estimate (done):  " + data.estimatedone);
-	console.log("Effort (total):   " + data.efforttotal);
-	console.log("Diff estimate:    " + (data.estimatedone - data.efforttotal));
-	console.log("----------");
-	console.log("");
-}
-
-