+++ run.js
@@ -10,6 +10,7 @@ var path = require('path');
 global.settings = require('./settings');
 settings.root   = __dirname.replace(/\/+$/, "");
 settings.exportPath = path.join(settings.root, 'export');
+settings.configPath = path.join(settings.root, 'config');
 settings.templatePath = path.join(settings.root, 'templates');
 settings.sprintTemplatePath = path.join(settings.root, 'templates' + path.sep + settings.template);
 settings.homeTemplatePath = path.join(settings.root, 'templates' + path.sep + settings.home_template);